+++ special-pages/pages/duckplayer/integration-tests/duckplayer-screenshots.spec.js
@@ -1,6 +1,6 @@
 /* global process */
 import { expect, test } from '@playwright/test';
-import { DuckPlayerPage } from './page-objects/duck-player.js';
+import { DuckPlayerPage } from './duck-player.js';
 
 test.describe('screenshots @screenshots', () => {
     test.skip(process.env.CI === 'true');
