+++ special-pages/pages/duckplayer/integration-tests/duckplayer-screenshots.spec.js-snapshots/regular-layout-macos-darwin.png
