+++ special-pages/pages/duckplayer/integration-tests/duckplayer-screenshots.spec.js-snapshots/tooltip-macos-darwin.png
