+++ special-pages/pages/duckplayer/integration-tests/duckplayer-screenshots.spec.js-snapshots/tooltip-windows-darwin.png
