+++ special-pages/pages/duckplayer/integration-tests/duckplayer-telemetry.spec.js
@@ -1,5 +1,5 @@
 import { test } from '@playwright/test';
-import { DuckPlayerPage } from './page-objects/duck-player.js';
+import { DuckPlayerPage } from './duck-player.js';
 
 test.describe('duckplayer telemetry', () => {
     test('sends a landscape impression', async ({ page }, workerInfo) => {
