+++ special-pages/shared/mocks.js
@@ -6,7 +6,7 @@ import {
     readOutgoingMessages,
     simulateSubscriptionMessage,
     waitForCallCount,
-} from '../../../messaging/lib/test-utils.mjs';
+} from '@duckduckgo/messaging/lib/test-utils.mjs';
 
 export class Mocks {
     /**
@@ -17,8 +17,8 @@ export class Mocks {
 
     /**
      * @param {import("@playwright/test").Page} page
-     * @param {import("../../../injected/integration-test/type-helpers.mjs").Build} build
-     * @param {import("../../../injected/integration-test/type-helpers.mjs").PlatformInfo} platform
+     * @param {import("injected/integration-test/type-helpers.mjs").Build} build
+     * @param {import("injected/integration-test/type-helpers.mjs").PlatformInfo} platform
      * @param {import("@duckduckgo/messaging").MessagingContext} messagingContext
      */
     constructor(page, build, platform, messagingContext) {
